requirements.txt
jiwer==3.0.4
pytest==7.3.2
black==23.3.0
pandas==2.2.2
httpx==0.27.0
pylint==2.17.4
fastapi==0.108.0
pytest-cov==4.1.0
pytest-timeout==2.3.1
pytest-asyncio==0.23.7
pytest-benchmark==4.0.0
websocket-client==1.8.0
python-multipart==0.0.9
openai-whisper==20231117
pylint-fail-under==0.3.0
uvicorn[standard]==0.30.1

#libs for sofya
faster-whisper==1.0.2
torchaudio
silero-vad
wave
python-Levenshtein
soundfile==0.12.1
hunspell==0.5.5
RapidFuzz==3.10.1